requirements.txt
apscheduler==3.11.2
attr==0.3.2
ConfigParser==7.2.0
cryptography==47.0.0
Cython==3.2.4
dl==0.1.0
docutils==0.22.4
fastapi==0.136.1
groq==1.2.0
HTMLParser==0.0.2
httpx==0.28.1
ipython==8.12.3
ipywidgets==8.1.8
Jinja2==3.1.6
jnius==1.1.0
keyring==25.7.0
langchain_community==0.4.1
langchain_core==1.3.2
langchain_groq==1.1.2
langchain_huggingface==1.2.2
lockfile==0.12.2
motor==3.7.1
numpy==2.4.4
pandas==3.0.2
passlib==1.7.4
Pillow==12.2.0
protobuf==7.34.1
pydantic==2.13.3
pymongo==4.17.0
pyOpenSSL==26.1.0
python-dotenv==1.2.2
python_jose==3.5.0
railroad==0.5.0
Sphinx==9.1.0
thread==2.0.6
tornado==6.5.5
trove_classifiers==2026.4.28.13
truststore
urllib3_secure_extra==0.1.0
uvicorn==0.46.0
xmlrpclib==1.0.1
